Revamped package and file structure
- placed all LIR Object types into lang/liro (LIR Object)
- split up aggregate files into individual files for each major class
- Lexer placed into lexer package
- parse nodes placed into parsetree package
- removed obsolete, unused files
- added file headers to all files missing headers
