+++ src/main/scala/org/cascas_project/cascas/lang/Context.scala
@@ -1,4 +1,14 @@
+//=============================================================================
+// lang/Context.scala : CaSCAS Project
+//=============================================================================
+
 package org.cascas_project.cascas.lang
+
+//=============================================================================
+
+import org.cascas_project.cascas.lang.liro.Identifier
+
+//=============================================================================
 
 case class Context (
   m: Map[Identifier, ContextValue] = Map[Identifier, ContextValue]()
