+++ src/main/scala/org/cascas_project/cascas/lang/Eval.scala
@@ -1,7 +1,13 @@
+//=============================================================================
+// lang/Eval.scala : CaSCAS Project
+//=============================================================================
+
 package org.cascas_project.cascas.lang
 
+//=============================================================================
+
 object Eval {
-  def apply(o: Object, ctx: Context): Object = {
+  def apply(o: liro.Object, ctx: Context): liro.Object = {
     o.eval(ctx).evaldObj
   }
 }
