+++ src/main/scala/org/cascas_project/cascas/lang/Evaluation.scala
@@ -1,5 +1,5 @@
 //=============================================================================
-// lang/TypedObject.scala : CaSCAS Project
+// lang/Evaluation.scala : CaSCAS Project
 //=============================================================================
 
 package org.cascas_project.cascas.lang
@@ -10,8 +10,8 @@ import org.cascas_project.cascas.lang.liro.Object
 
 //=============================================================================
 
-case class TypedObject(tpe: TypeIdentifier, value: Object) extends ContextValue {
-
-  override def toString(): String = f"$tpe -> $value"
-
+case class Evaluation(evaldObj: Object, ctxDelta: ContextMutationSet) {
+  def keepOnlyReassignments(): Evaluation = {
+    Evaluation(evaldObj, ctxDelta.onlyReassignments())
+  }
 }
