--- src/main/scala/org/cascas_project/cascas/lang/Literal.scala
@@ -1,3 +1,0 @@
-package org.cascas_project.cascas.lang
-
-trait Literal extends Object {}
