--- src/main/scala/org/cascas_project/cascas/lang/Object.scala
@@ -1,8 +1,0 @@
-package org.cascas_project.cascas.lang
-
-trait Object {
-  def eval(ctx: Context): Evaluation
-  def checkType(ctx: Context, tpe: TypeIdentifier): Boolean
-  def inferType(ctx: Context): Option[TypeIdentifier]
-  def toRepr(): String = this.toString()
-}
