+++ src/main/scala/org/cascas_project/cascas/lang/OperatorType.scala
@@ -1,4 +1,14 @@
+//=============================================================================
+// lang/OperatorType.scala : CaSCAS Project
+//=============================================================================
+
 package org.cascas_project.cascas.lang
+
+//=============================================================================
+
+import org.cascas_project.cascas.lang.liro.Identifier
+
+//=============================================================================
 
 case class OperatorType(
   args: Vector[FormalParameter],
@@ -47,22 +57,3 @@ object OperatorType {
   }
 
 }
-
-
-
-
-case class FormalParameter(id: Identifier, tpe: TypeIdentifier, manyness: Manyness = One)
-
-
-
-
-import org.cascas_project.cascas.shared.Enumerated
-
-sealed abstract class Manyness extends Enumerated.Value[Manyness] {
-  def enumType = Manyness
-}
-final object Manyness extends Enumerated.Type[Manyness] {
-  val enum = Vector(One, Many)
-}
-final case object One extends Manyness {}
-final case object Many extends Manyness {}
