+++ src/main/scala/org/cascas_project/cascas/lang/TypeIdentifier.scala
@@ -1,4 +1,10 @@
+//=============================================================================
+// lang/TypeIdentifier.scala : CaSCAS Project
+//=============================================================================
+
 package org.cascas_project.cascas.lang
+
+//=============================================================================
 
 trait TypeIdentifier extends ContextValue {}
