+++ src/main/scala/org/cascas_project/cascas/lang/liro/Expr.scala
@@ -1,11 +1,9 @@
 //=============================================================================
-// lang/package.scala : CaSCAS Project
+// lang/liro/Expr.scala : CaSCAS Project
 //=============================================================================
 
-package org.cascas_project.cascas
+package org.cascas_project.cascas.lang.liro
 
 //=============================================================================
 
-package object lang {
-
-}
+trait Expr extends Object {}
