+++ src/main/scala/org/cascas_project/cascas/lang/liro/IfExpr.scala
@@ -1,0 +1,17 @@
+//=============================================================================
+// lang/liro/WhileExpr.scala : CaSCAS Project
+//=============================================================================
+
+//package org.cascas_project.cascas.lang.liro
+
+//=============================================================================
+
+import org.cascas_project.cascas.lang.Context
+import org.cascas_project.cascas.lang.ContextMutationSet
+import org.cascas_project.cascas.lang.TypeIdentifier
+
+//=============================================================================
+
+//case class IfExpr(ifdo: Vector[TestAndBody], elsedo: Object) extends Expr {
+//
+//}
