+++ src/main/scala/org/cascas_project/cascas/parser/Item.scala
@@ -1,7 +1,10 @@
 //=============================================================================
-// Item.scala : CaSCAS Project
+// parser/Item.scala : CaSCAS Project
 //=============================================================================
+
 package org.cascas_project.cascas.parser
+
+//=============================================================================
 
 case class Item(
   val lhs: Symbol,
