+++ src/main/scala/org/cascas_project/cascas/parser/package.scala
@@ -1,7 +1,12 @@
+//=============================================================================
+// parser/package.scala : CaSCAS Project
+//=============================================================================
+
 package org.cascas_project.cascas
 
+//=============================================================================
 // This package contains modules used in parsing.
-
+//
 package object parser {
 
   private[parser] type CFGRule = (Symbol, Vector[Symbol])
